pyproject.toml
[tool.poetry]
name = "django-extra"
version = "0.1.0"
description = ""
authors = ["Adilet <119230294+siweirdb@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
